core: use slice of u32 instead of i32 for data of TLV
This commit is a preparation to use alsa-ctl-tlv-codec crate to build
data of TLV. The alsactl crate has a bug to accept array of i32 element,
thus current code include structures to have static array of i32 element.
They should be u32 array to use the crate. Instead, methods of CardCntr
structure are written to convert these arrays.

+++ libs/bebob/src/maudio/profirelightbridge_model.rs
@@ -133,7 +133,7 @@ impl<'a> MeterCtl {
     const METER_MIN: i32 = 0;
     const METER_MAX: i32 = 0x007fffff;
     const METER_STEP: i32 = 256;
-    const METER_TLV: &'a [i32] = &[5, 8, -14400, 0];
+    const METER_TLV: &'a [u32] = &[5, 8, -14400i32 as u32, 0];
 
     fn new() -> Self {
         MeterCtl {
